fix: split OpenCode agent tests to satisfy backend size limits
Move OpenCode-agent regressions into dedicated test files so
interfaces_contract_test.go, factory_test_helpers_test.go, and
agent_test.go stay within the 1000-line backendsizecheck limit.
