+++ docs/internal/processes/factory-workstation-relevant-files.md
@@ -39,8 +39,9 @@ work inputs.
 - When the public workstation outcome contract changes, update the live prompts, `pkg/cli/init` scaffold text, and the workstation authoring/reference docs together so generated workflow text does not drift from the checked-in maintainer flow.
 - Keep backend runner precedence in `pkg/interfaces.ResolveRunnerSelection(...)` and let service-layer workstation validation key only off `ResolvedRunnerSelection.Source`; do not re-derive workstation, factory, or legacy-provider precedence from raw config fields in `pkg/service`.
 - Keep OpenCode agent precedence in `pkg/interfaces.ResolveOpenCodeAgent(...)` with worker/workstation `openCodeAgent` parsed through `pkg/config/agents_config.go` and merged in `pkg/config/runtime_config.go`; later dispatch stories should consume the resolved value instead of re-reading frontmatter.
-- Reject non-empty `openCodeAgent` when `interfaces.ResolveRunnerSelection(...)` does not resolve to `opencode` via `interfaces.ValidateOpenCodeAgentForRunnerSelection(...)` from `pkg/service/factory_build.go` `validateConfiguredWorkstationRunners`; prove mismatches with `TestRunnerSelectionValidation_RejectsOpenCodeAgentOnNonOpenCodeRunner` and `TestLoadWorkersFromConfig_RejectsOpenCodeAgentOnNonOpenCodeRunner`.
-- Populate `interfaces.ProviderInferenceRequest.OpenCodeAgent` in `pkg/workers/executor/agent.go` `inferenceRequestForExecutionRequest` only when the effective model provider is `opencode`, using `interfaces.ResolveOpenCodeAgent(...)` from runtime workstation and worker config; prove forwarding with `TestInferenceRequestForExecutionRequest_SetsOpenCodeAgentForOpenCodeDispatch`, `TestInferenceRequestForExecutionRequest_LeavesOpenCodeAgentEmptyForNonOpenCodeDispatch`, and `TestAgentExecutor_ForwardsOpenCodeAgentOnOpenCodeDispatch`.
+- Reject non-empty `openCodeAgent` when `interfaces.ResolveRunnerSelection(...)` does not resolve to `opencode` via `interfaces.ValidateOpenCodeAgentForRunnerSelection(...)` from `pkg/service/factory_build.go` `validateConfiguredWorkstationRunners`; prove mismatches in `pkg/service/factory_opencode_agent_test.go` (`TestRunnerSelectionValidation_RejectsOpenCodeAgentOnNonOpenCodeRunner`, `TestLoadWorkersFromConfig_RejectsOpenCodeAgentOnNonOpenCodeRunner`).
+- Keep OpenCode-agent contract tests in dedicated files when parent packages approach the 1000-line backend-size limit: `pkg/interfaces/opencode_agent_contract_test.go`, `pkg/service/factory_opencode_agent_test.go`, and `pkg/workers/executor/agent_opencode_agent_test.go`.
+- Populate `interfaces.ProviderInferenceRequest.OpenCodeAgent` in `pkg/workers/executor/agent.go` `inferenceRequestForExecutionRequest` only when the effective model provider is `opencode`, using `interfaces.ResolveOpenCodeAgent(...)` from runtime workstation and worker config; prove forwarding in `pkg/workers/executor/agent_opencode_agent_test.go`.
 - Append `opencode run --agent <name>` in `pkg/workers/provider/provider_behavior.go` `openCodeProviderBehavior.BuildArgs` when `req.OpenCodeAgent` is non-empty, placing `--agent` after `--model` and before `--session`; prove with `TestOpenCodeProviderBehavior_BuildArgs` cases `WithOpenCodeAgent` and `WithOpenCodeAgentModelSessionWorkingDirectoryAndSkipPermissions`.
 - Record configured OpenCode agents in provider request metadata as `opencode_agent` from `workDiagnosticsForInferenceRequest` in `pkg/workers/executor/diagnostics.go` and `pkg/workers/provider/helpers.go` only when `ProviderInferenceRequest.OpenCodeAgent` is non-empty; allowlist the key in `pkg/interfaces/safe_diagnostics.go` `isSafeProviderMetadataKey` and prove with `pkg/workers/provider/diagnostics_test.go` plus `TestAgentExecutor_ForwardsOpenCodeAgentOnOpenCodeDispatch`.
 - Extend public factory OpenAPI `Worker` and `Workstation` schemas with optional `openCodeAgent`, regenerate Go/TypeScript clients, map through `pkg/config/factory_config_mapping.go` and `factory_config_mapping_internal.go` via `openCodeAgentInternalFromAPI`, and prove decode/projection with `pkg/config/openapitests/openapi_factory_opencode_agent_roundtrip_test.go` plus `ui/src/api/factory-definition/api.test.ts` non-empty validation.
